# lib/api-zod/src/generated/types/activityItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sky-Hosting API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ActivityItem {
  id: string;
  type: string;
  message: string;
  userId: string;
  username: string;
  /** @nullable */
  deploymentId?: string | null;
  /** @nullable */
  deploymentName?: string | null;
  createdAt: Date;
}