# lib/api-zod/src/generated/types/systemSettings.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sky-Hosting API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SystemSettings {
  baseDomain: string;
  maxDeploymentsPerUser: number;
  allowPublicRegistration: boolean;
  defaultBuildTimeout: number;
  enableDockerDeployments: boolean;
  maintenanceMode: boolean;
  rateLimitPerMinute: number;
}