# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sky-Hosting API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from "./userRole";

export interface User {
  id: string;
  username: string;
  /** @nullable */
  email?: string | null;
  role: UserRole;
  isActive: boolean;
  /** @nullable */
  apiKeyPrefix?: string | null;
  deploymentCount: number;
  createdAt: Date;
  updatedAt: Date;
}